Revert "Add support for getting/tailing deploy logs (#762)"
This reverts commit ac123daf7c107d835968973e55ef6b8639aecd69.
