+++ package.json
@@ -94,7 +94,6 @@
     "express": "^4.17.1",
     "express-logging": "^1.1.1",
     "find-up": "^3.0.0",
-    "firebase": "^5.11.1",
     "fs-extra": "^8.1.0",
     "fuzzy": "^0.1.3",
     "get-port": "^5.1.0",
@@ -106,7 +105,6 @@
     "inquirer": "^6.5.1",
     "inquirer-autocomplete-prompt": "^1.0.1",
     "is-docker": "^1.1.0",
-    "isomorphic-ws": "^4.0.1",
     "jwt-decode": "^2.2.0",
     "lambda-local": "^1.7.1",
     "lodash.get": "^4.4.2",
@@ -142,8 +140,7 @@
     "wait-port": "^0.2.2",
     "winston": "^3.2.1",
     "wrap-ansi": "^6.0.0",
-    "write-file-atomic": "^3.0.0",
-    "ws": "^7.2.3"
+    "write-file-atomic": "^3.0.0"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1.22.2",
