+++ src/commands/deploy.js
@@ -13,8 +13,7 @@ const randomItem = require('random-item')
 const inquirer = require('inquirer')
 const SitesCreateCommand = require('./sites/create')
 const LinkCommand = require('./link')
-const { NETLIFYDEVLOG } = require('../utils/logo')
-const { waitForBuildFinish } = require('../utils/logs.js')
+const { NETLIFYDEV } = require('../utils/logo')
 
 class DeployCommand extends Command {
   async run() {
@@ -23,8 +22,6 @@ class DeployCommand extends Command {
 
     const deployToProduction = flags.prod
     await this.authenticate(flags.auth)
-
-    const [accessToken] = this.getConfigToken()
 
     await this.config.runHook('analytics', {
       eventName: 'command',
@@ -81,20 +78,15 @@ class DeployCommand extends Command {
     if (flags.trigger) {
       try {
         const siteBuild = await api.createSiteBuild({ siteId: siteId })
-        this.log(`${NETLIFYDEVLOG} A new deployment was triggered successfully.`)
-        this.log(`${NETLIFYDEVLOG} Waiting for the build to start ⏳`)
-        const deployData = await api.getDeploy({ deployId: siteBuild.deploy_id })
-        if (deployData && deployData.log_access_attributes && Object.keys(deployData.log_access_attributes).length) {
-          await waitForBuildFinish(api, site.id, siteBuild.deploy_id, accessToken)
-        }
+        this.log(`${NETLIFYDEV} A new deployment was triggered successfully. Visit https://app.netlify.com/sites/${siteData.name}/deploys/${siteBuild.deploy_id} to see the logs.`)
+        return
       } catch (err) {
         if (err.status === 404) {
           this.error('Site not found. Please rerun "netlify link" and make sure that your site has CI configured.')
         } else {
-          this.error(err)
+          this.error(err.message)
         }
       }
-      this.exit()
     }
 
     // TODO: abstract settings lookup
