--- src/utils/log_streamers/resources.js
@@ -1,6 +1,0 @@
-const LogResources = {
-  BUILD: "build",
-  FUNCTION: "function"
-};
-
-module.exports = LogResources;
